Get byte tests and such

+++ include/ceebee/cpu.h
@@ -46,7 +46,7 @@ unsigned short* getRegister16(CPU *cpu, int index);
 unsigned char* getRegister(CPU *cpu, int index);
 
 unsigned char getByte(unsigned char const *cart, unsigned short addr);
-unsigned int getNN(unsigned char const *cart, unsigned short addr);
+unsigned short getNN(unsigned char const *cart, unsigned short addr);
 
 void printCpu(CPU cpu);
 void run_cycle(CPU *cpu, unsigned char const *cart);
